+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## v4.20.0
+  - [EE] Support batch processing of multiple documents (max 5000) with one API call. The result is a zip file of documents.
+    This feature is temporarily limited to some customers. Please contact us if you need to use it.
+    Rate-limit : one batch processing at a time by tenant.
+
 ## v4.19.0
   - Release February 22th 2024
   - [EE] The `:color` formatter supports updating the background color of a page with `:color(page, background)` options in DOCX templates.